split routeguide server
